# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PantryPal API",
    "description": "PantryPal inventory management API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/subscription/checkout": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "summary": "Checkout",
        "operationId": "checkout_api_subscription_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subscription/portal": {
      "post": {
        "tags": [
          "Subscription"
        ],
        "summary": "Portal",
        "operationId": "portal_api_subscription_portal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutRequest": {
        "properties": {
          "tier": {
            "type": "string",
            "enum": [
              "pro",
              "family"
            ],
            "title": "Tier"
          },
          "billingInterval": {
            "type": "string",
            "enum": [
              "month",
              "year"
            ],
            "title": "Billinginterval"
          },
          "successUrl": {
            "type": "string",
            "title": "Successurl"
          },
          "cancelUrl": {
            "type": "string",
            "title": "Cancelurl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tier",
          "billingInterval",
          "successUrl",
          "cancelUrl"
        ],
        "title": "CheckoutRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PortalRequest": {
        "properties": {
          "returnUrl": {
            "type": "string",
            "title": "Returnurl"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "returnUrl"
        ],
        "title": "PortalRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}